+++ src/frontend/komponenter/Fagsak/Behandlingsresultat/Kompetanse/KompetanseSkjema.tsx
@@ -6,8 +6,8 @@ import { Alert, Heading } from '@navikt/ds-react';
 import type { FeltState } from '@navikt/familie-skjema';
 
 import type { IBehandling } from '../../../../typer/behandling';
-import type { IKompetanse } from '../../../../typer/kompetanse';
-import { KompetanseStatus } from '../../../../typer/kompetanse';
+import type { IKompetanse } from '../../../../typer/eøsPerioder';
+import { KompetanseStatus } from '../../../../typer/eøsPerioder';
 import KompetanseTabellRad from './KompetanseTabellRad';
 
 const KompetanseContainer = styled.div`
