+++ src/frontend/typer/eøsPerioder.ts
